Implement initial version of a hypervisor ram log
Ramlog allows to store data that reveals a hypervisor work

As it could contain crucial information ramlog entries being encrypted
For that purpose the log logically divied on 64byte size entries to be
usable with Chacha20 stream cipher algorithm which was chosen

In this stage keys are dummy ones and stored directly in sources, its so
before keystorage will be designed and implemented.

To track log entries in time timestamp generation was implemented, it
is not corresponding to kernel timestamps. To distinguish them in
possible shared output hyp ram log timestamps marked as 'rl'

Signed-off-by: Grigoriy Romanov <grigoriy.romanov@unikie.com>