HACK: Workaround to get "protected mode with firmware" to work
"protected with firmware" mode requires that the SMMU is initialized.
This commit allows the mode to be used without the SMMU.
In "protected with firmware" mode, MMIO GUARD is enabled.
For some reason, the guest will not boot if it is enabled.
his commit removesthe MMIO GUARD enablement

NOTE: This commit removes important security tests and should only
be used for testing purposes

Signed-off-by: Martti Takala <martti.takala@unikie.com>